Add vtkCellQuality which evaluates various qualities for cells in a mesh.
vtkCellQuality evaluates various qualities for the cells in a mesh. The
name is chosen following vtkMeshQuality which behaves similarly to this
filter with some behaviors that is undesired. vtkCellQuality improves
vtkMeshQuality by returning reasonable invalid values for cases that it
does not support. Specifically, UndefinedQuality is returned for any
query of quality that is not well defined for the type of the cell, e.g.
query for the volume of a triangle and UnsupportedGeometry is returned
for any type of geometry that is not supported by the filter. The
invalid values can be set by the caller before calling the filter which
gives more reasonable results. For a list of supported qualities and
geometries, please check vtkCellQuality.h