Add SetVariantValue method to vtkAbstractArray and subclasses
We have InsertVariantValue, which is analogous to InsertValue in that
it's SetValue with bounds checking, but no SetVariantValue.  I've added
this to vtkAbstractArray and subclasses and modified InsertVariantValue
to do the bounds check and fall through to SetValue.  InsertVariantValue
is still virtual so you can override it if you have a very good reason
but that probably won't be necessary.