ENH: Add a preprocessor to the wrapper's parser.
The preprocessor allows robust handling of if, define, and include
directives.  It contains its own simple parser for evaluating
expressions using integer math.  If given include paths via "-I"
it searches those paths, but if it does not find an include file
then it continues without failing.  Included files are only read
by the preprocessor, not by the main parser.