Fix reporting netCDF variables for each dimension set in ParaView.
The object inspector for the netCDF reader was reporting the incorrect
variables for each dimension set.  The issue was that the
ArraySelectionInformationHelper was sorting the variables in
VariableArrayInfo so that they did not match up with the
VariableDimensionsInfo.  Fixed the problem by creating a new method
to vtkNetCDFReader that returns the variable array names in a
vtkStringArray that can be loaded by ParaView without sorting.