BUG: LIC algorithms are working again on Linux.
vtkLineIntegralConvolution2D_fs1.glsl was using bReset in an expression to
compute tcord0. As it was not a conditional, there was a texture fetch on a
uninitialized texture. The result could be NaN or undefined and this value
multiplied by (1-bReset) would lead to a NaN or undefined result that would
propagate to the final value of tcord0 even if the actual value of (1-bReset)
is 0. The fix is to use a conditional statement instead.