ENH: New vtkFreeTypeTools class.
Added a new vtkFreeTypeTools class. This is intended to be a low level
class to use the FreeType2 library to render strings to images. It
borrows a lot of code from vtkFreeTypeUtilities, but uses a simpler
approach and gets rid of a lot of the old API.

Initially this class has been implemented with the 2D API in mind. It
will be used to provide a low level interface to font rendering. It is
able to render Unicode characters, but the embedded fonts do not contain
any Unicode characters. The hard coded path will need to be replaced
with a more general mechanism for loading fonts from files.