BUG: Undo a bug fix that did bad data object  manipulation.
This removes one of the fixes for Bug 1452 that involved directly
setting the Spacing and Origin upstream vtkImageStencilData objects
rather than acting through the pipeline.  Since that bug fix,
stencil sources have been given an InformationInput so that the
Spacing and Origin can be transferred to the vtkImageStencilData
via the pipeline.