BUG: Further fixes for bug 11550, more tolerance.
The test was failing for a couple machines, my best guess
is that extended precision on the x87 FPU was to blame.
This patch always uses the Bounds ivar instead of recomputing
the bounds (to ensure numerical consistency), and always
includes a small tolerance across 2D images to avoid the
need for strict floating-point equality just in order to
get a hit.