ENH: Introduce AMR-Kit & Base/Utility Classes
Introduce AMR-Kit in to the build and two additional
classes: (a) vtkSerializable and (b) vtkAssertUtils.
The AMR-Kit is geared towards encapsulating all of
the AMR infrastructure & algorithms. vktSerializable
defines a pure abstract interface for all serializable
objects. Last, the vtkAssertUtils is a singleton class
provides functionality for self-checking that is enabled
only when -DASSERT_ON is provided.