ENH: Part 3 of the multi-threading and optimization of space leaping
This adds customized functions for traverising the input scalars in
one pass to compute the min-max-gradient_max space leaping structure.
Caching of the output to ensure that the entire output is not regenerated
based on flags is also added. It was missing earlier.

Please see the earlier two feature commits :

  MultiThread-SpaceLeaping-VolumeReadering-Feature*