BUG 0011837: Fix rare incorrect warp results for vtkImageReslice.
If the reslice input origin, spacing are 0.0, 1.0 and if the ResliceAxes
are not set or are identity, then any warp transform (i.e. transform not
expressible as a 4x4 matrix) will be applied with an incorrect shift or
scale applied to the coordinates, unless the output origin, spacing are
also 0.0, 1.0.