vtkTable allowed you to access off the end
vtkTable did not check that the requested data item actually
existed and returned bogus data off the end of a query.  I used
the same error return used elsewhere for that case.

Latest change uses the number of tuples instead of the maximum id
to determine when you are off the end.

Change-Id: I1be5c414048c90f7f7444a9c2363b17b8a2740f7