ENH: Interlevel point ownership implementation
Separate the point ownership information calculation
in two separate routines: (1) CheckOwnershipAtSameLevel()
and (2) CheckOwnershipDownstream(). The first computes the
point ownership information within the same level. In this
case, points from different blocks within the same level
can fall on top of each other, the block with the smallest
block ID wins and acquires ownership of the point. In the
latter case, all points that fall within a hi-resolution
block are not owned by the grid.