ENH: Additional logic for constructing dual cells
Implemented additional logic for constructing dual
cells. The approach consists of two steps. First,
vtkAMRDualMeshExtractor::ProcessCellDual() determines
if a dual cell can be formed from the given dual node
based on: (a) if the cell is visible on the input grid
and (b) if it owns one or more of its points. Second,
the vtkAMRDualMeshExtractor::GetCellIds() method checks
if a valid cell can be formed by: (a) checking if the
resulting cell will be within the cell index bounds of
the input grid and (b) by checking the visibility of the
dual cell nodes of the resulting cell. If all the dual cell
nodes correspond to cells in the input grid that are not
visible the cell is rejected, o/w, it is accepted.