ENH: Make Extract dataset output to be Multiblock
Changed vtkExtractDataSets s.t. the output is a an
instance of vtkMultiBlockDataSet. Each block in the
multiblock dataset contains a vtkMultiPieceDataSet instance.
Hence, a block in the output corresponds to a level in the
input AMR vtkHierarchicalBoxDataSet. This change was necessary
b/c of the way vtkHierarchicalBoxDataSet is rendered. Recent
changes imposed that only boundary blocks are processed and
further, only block faces on the actual boundary are rendered.
Consequently, the way vtkExtractDataSets was previously implemented,
i.e., by extracting the data in a new vtkHierarchicalBox instance,
was resulting in faulty behavior selected blocks that are not on
the boundary were not rendered/ignored by the vtkPVGeometry
and vtkDataSetSurface filters.