BUG: The vtkImageSliceMapper didn't notice UpdateExtent changes.
If the input to the vtkImageSliceMapper had multiple consumers,
and one of those consumer set the input's UpdateExtent to something
other than the UpdateExtent needed by the mapper, the pipeline did
not know that it had to update and would crash as a result.  Adding
a "fake" output to the mapper works as a fix.