Add point edition in vtkControlPointsItems
-You can edit control points in 2 fashions:
 If vtkControlPointsItem::StrokeMode is true, you can "draw" points while
the left mouse button is down, it will add control points under the mouse
cursor.
 If vtkControlPointsItem::StrokeMode is false, you can drag&drop existing
points. SwitchPointsMode control the behavior to follow when a point
crosses another.
In both modes, if you click the left button mouse not on a point, a point is
created under the mouse cursor.
-Add vtkCompositeControlPointsItem to mix ct function and opacity function.
It creates points for Color Transfer and Opacity
functions.Both functions will share the same control points.
-Add vtkControlPointsItem::EditPoint to control midpoint and sharpness
-Support null transfer functions in the ScalarsToColors items