QVTKWidget was capturing image from the buffer at wrong time.
SwapBuffer behaves differently on different platforms. On some it seems to paste
contents of back buffer to front while on others it literally swaps them. The
QVTKWidget needed to capture back buffer contents while caching to avoid issues
with overlapped windows. So it needed an event from vtkRenderWindow that was
fired after render was done but before swapping of the buffers. So changed
vtkRenderWindow to fire such an event and QVTKWidget to capture the buffer on
that event.