Fixed BUG #12202. vtkXMLUnstructuredGridReader memory blot.
vtkXMLUnstructuredGridReader kept on increasing memory size as one when over
timesteps for a temporal dataset. The issue was that the reader was
miscalculating the offset into the cellarray for the active piece. That offset
kept on increasing for every timestep. Fixed that.

Also fixed vtkXMLDataReader to mark array modified after reading so that they
realize that the internal data has been changed.

Change-Id: I0e6f196d3da41af1ec0337cf86cafa55e35e8635