Started removing data object's dependency on the pipeline.
It was decided to remove any dependencies that data objects
have on the pipeline logic. When modularization is complete,
this will allow us to build a small "data model" library
that does not depend on the "execution model". It also
cleans up a lot of the interdepencies between data objects
and pipeline code. To achieve this, we need to remove all
functionality that depend on executives and pipeline logic
from vtkDataObject and subclasses. This includes any meta-data
such as whole extent as well as methods to setup pipeline
connectivity such as SetInput (to be removed from algorithms).