Bug fixes for vtkTensorGlyph
1) fix orientation inversion causing flipped surface normals. When the
determinant of the transformation matrix is < 0 surface normals are
flipped. The patch fizes it by testing for this condition and applying
an additional scale(-1,-1,-1) to the surface normal transform, leaving
the surface inverted but with outward pointing normals.

2) Added names to generated output arrays so that they are recongnized
by ParaView.

3) Override RequestUpdateExtent so that glyph source generates an whole
glyph when running in parallel, with out this only a piece of the glyph
is generated.

Change-Id: Iad3920dd623b80bb7fe8a8a924aeb2c42a830e86