Using a vtkWeakPointer to avoid a dangling pointer.
VTK_GLOBAL_MULTI_PROCESS_CONTROLLER had static Set/Get
methods but never did referenced the variable so it
wasn't guaranteed that once it was set that the
instance wasn't deleted.  Changed to a weak pointer
so that it would be set to NULL if it was deleted
while still not modifying the reference count.

Change-Id: I218dd6b128754beceb75a935b31bd1b989008216