BUGFIX: Fix issue with acquiring a reference grid
Fixed issue when acquiring a reference grid. By default,
the code was looking at block (0,0) to determine which
arrays are loaded etc. However, when the AMR dataset is
distributed, not all processes will have block(0,0).To resolve
this issue the code loops through the AMR hierarchy until a
non-null dataset is retrieved. That dataset is then used
as the reference grid.