ENH: Widget to delineate 27 clique ROIs during volume rendering
This widget displays a set of axis aligned lines that can be interactively
manipulated to crop a volume. The region to be cropped away is displayed
in a different highlight. Much like the vtkVolumeMapper, this widget
supports 27 possible configurations of cropping planes. (See
CroppingRegionFlags). If a volume mapper is set, the cropping planes
are directly propagated to the volume mapper. The widget invokes a
CroppingPlanesPositionChangedEvent when the position of any of the
cropping planes is changed.