BUGFIX: Parsing of cfactors before filename is set
Fixed issue wherein the method vtkAMREnzoReader::ParseConverionFactors
was called before this->FileName was set. This was causing an assert
failure. ParseConversionFactors is now properly called. In addition,
added logic to handle the case where an ENZO conversion parameters
file is missing or otherwise not available to improve the robustness
and flexibility of the reader.