Adding vtkArrayData reader and writer.
vtkArrayDataWriter uses vtkArrayWriter to serialize the collection
of arrays stored in vtkArrayData. vtkArrayDataReader similarly uses
vtkArrayReader to read in vtkArrayData instances. vtkArrayReader
needed to be modified slightly so that it no longer requires that
the stream ends when the last value in the array is read, since when
reading vtkArrayData instances with multiple arrays this will not be
the case. Because of this change I needed to remove the test code
that tested that there is an error if the stream continues past the
last array value. TestArrayDataWriter tests the writer and the reader
to ensure the same data is present after the round trip.