Force the opengl buffer to be fully refreshed.
A hack in QWidgetPrivate::repaint_sys (qbackingstore.cpp) checks the type
of the widget QPaintEngine. If it is an OpenGL engine, it then forces the
 repaint to be done on the entire widget instead of the dirty region.

Changing the paint engine type, ensures QVTKWidget to be fully repainted on Windows when resizing the widget.
This fix makes the behavior of QVTKWidget similar to QVTKWidget2.
It has visible results when the flag Qt::WA_PaintOnScreen is set on Windows (not the case because of the previous commit)

Change-Id: I5ce944d73c5931ca9423b9cc848bc1ede6bf5b91