ENH: Enhancements to vtkTulipReader and new support for clusters.
1. Fixed previous behaviour whereby properties were assumed to be on
both nodes and edges, leading to empty arrays being added to the
graph when properties are only set for either node or edge.

2. Node and edge pedigree ids are added that retain the id values given in
the input Tulip file.

3. Added double property type.

4. Updated vtkTulipReaderTestProperties to cover 1, 2, and 3 above.

4. Added ability to read clusters.
Clusters (subgraphs) are output as vtkSelectionNodes within the
vtkAnnotationsLayer on output port 1. Nested hierarchies are
flattened.

5. Added a filter to convert a vtkPolygon to a polygon of the convex hull
of the points in the input. The filter ensures visibility by adding extra
points and/or scaling the hull up. Static convenience methods provide a
similar function, with vtkPoints as the input/output. Adding a vtkRenderer
allows a minimum size in pixels to be specified.

6. Added a filter to display graph annotations.
The new filter takes a vtkAnnotations layers containing selections
of vertices from a vtkGraph. Each node of the selection is
represented by it's convex hulli/bounding rectangle.

6. Added a test (TestTulipReaderClusters) that uses new
functionality in vtkTulipReader to read in and display
clusters/annotations, covering 4, 5 and 6 above.

7. Added a test (TestConvexHull2D) to exercise the static methods
in vtkConvexHull2D.

Change-Id: I424a0a02e9ac82fdce78d931819f1dffce25796e