BUG: Fixed issue with font caching in charts.
The equality operator in the font caching for the charts was evaluating
to true even when the font size and/or color had been changed. This
caused several issues on subsequent renders where the font metric did
not match the rendered string and changes were not visible. Added a new
test that loads the string cache with a couple of strings, and then
changes properties.

I am not certain I covered every change, if the unique id for text
properties might be improved further and whether the extra comparisons
necessary the performance significantly (especially for smaller
strings).

Change-Id: Ie9d3a3e7903c9cd3ef122b7197f491d04434e122