ENH: Structured Grid Connectivity
Adding vtkStructuredGridConnectivity singleton class that implements
methods for computing the inter-grid connectivities based on:
(a) the whole extent.
(b) the global grid extent of each grid.

The neighboring information for each neighbor of a grid are stored in
the vtkStructuredNeighbor data-structure, an internal light-weight
object, which stores the neighbor grid ID and the overlap extent.

Change-Id: I14da13d0696cd328485eead60b9d44f6cac5eb3f