BUG: Use of uninitialised value of size 8
This valgrind defect has been present for years.

There are instances where vtkDoubleScreenEdge::Case is not defined
before RasterizeTriangle. The logic in vtkSimpleScreenEdge::Init is
complex. Perhaps in the future, a more more extensive examination will
uncover the real problem. For now, the patch fixes the valgrind defect
and still passes the regression test.

Change-Id: If8a9846d1a4561447d39078daa70c5acebd8f394