ENH: Improved Resample Filter Speed
If a previous grid does not contain the current point then the
level containing the grid is tested.  If the level does contain the
point the newly found grid is used that and the next level is tested
Else the level is reset to 0 and the max level is set to 1 below
the previous level.

Also added a trivial bounds test.

Also fixed a bug that caused the resulting uniform grid to be
larger than the requested resolution.

Change-Id: I83656837ae2cf0e63d96a3f00d4db06ecc6ff3e7