ENH: Changed AddExtent to take just the extent
Changed the signature of AddExtent to take just the extent
to be added, i.e., AddExtent( int ext[6] ), instead of
AddExtent( const int idx, int ext[6] ). Since the extent is
now always added to the end of the list, there is no need to
pass an index. Previously, the partition extents list was
pre-allocated in which case passing the index made sense.

Change-Id: I90c05a9f4fc6810323fa0dcd4e3f85f31913d85d