ENH: Enable re-use of extent partioner instance
This commit introduces changes necessary for reusing an extent
partitioner instance to partition another extent and/or to generate
a different number of partitions. When the user calls SetNumberOfPartitions
or SetGlobalExtent, the previously computed sub-extents are cleared and
the state of the partitioner is re-initialized. In addition, the GlobalExtent
array is now ordered as [imin,imax,jmin,jmax,kmin,kmax] which is compliant to
the VTK extent convention.

Change-Id: I65ae6721f98003cedab72803e71dce95d8fae7c4