COMP: Ignoring return value
Some newer gcc compilers warn:

warning: ignoring return value of 'XXX', declared with attribute
warn_unused_result

Regarding vtkWindBladeReader: vtkWarningMacro is used to report the
errors rather than vtkErrorMacro.
As noted by Dave DeMarle in this gerrit review: The only problem with
the change is that it exposes deeper problems in the reader. The
deeper problems shouldn't prevent the change from going in. While we
wait for the deeper fixes, wouldn't something like this be less
invasive? It would fix the compiler warnings, remind us developers to
fix it, and not make the behavior any worse than it is now.

Change-Id: I6cd25007a574346307bad884ef47582740207abb