BUG: Set number of outputs and other changes.
The constructor of vtkCenterOfMass was not setting the number of
outputs to zero.  Also, having the static method take a vtkDataSet
as input was not flexible... the caller should have the freedom
of providing their own points array and scalar array.  A further
change was the addition of asserts, which seem to be the way that
VTK developers have been adding error checking to static methods
recently.

Change-Id: Ia61c6479064556c0086de4bb726bda5bd316676a