ENH: Define BlockFace enum for BlockTopology
BlockFace is used to identify the 6 block faces.
The enum is used to: (1) define a block's connectivity
connections, e.g., a block has a connection @BlockFace::RIGHT
which corresponds to the +i direction, and (2), to identify
a node's orientation on the boundary of the block. For example,
given the i-j-k coordinates of the node, it can be determined
on which face of the block the node is in, if it is on a block
boundary.

Change-Id: I4a8b8d40d9a27a0ce46e06bdd053e5b46d3c1421