ENH: A b-spline image interpolator and a b-spline warp transform.
This commit contains code that originated in the VTK Journal:
DG Gobbi and YP Starreveld
"Uniform B-Splines for the VTK Imaging Pipeline"
http://hdl.handle.net/10380/3252

The vtkImageBSplineInterpolator is an image interpolator that uses
uniform b-splines.  Prior to interpolation, images must be filtered
with vtkImageBSplineCoefficients to compute the knot coefficients.
The vtkBSplineTransform implements a cubic b-spline warp grid.

Change-Id: Ib8c95cc736fc7db194cc1bcd21ef48c1e82e16a2