ENH: Added field comparisson test
Added test to ensure that the ghosted grid has correctly transferred the
the data from the neighboring grids. The test relies on the fact that the
initial field that is applied is based on the cell's centroid and node
coordinates. Hence, applying the same function on the ghosted dataset yields
the expected results that are compared with the computed data.

Change-Id: Idbcd4bcd33184d29887a059d59f75d45c8b8c072