ENH: Change interpolator Mirror feature to mirror at boundary.
The SetBorderModeToMirror causes indices to be mirrored when
out-of-bounds lookups are done.  The "mirror" used to be half
a pixel beyond the center of the edge pixels (like OpenGL),
which led to duplication of the edge pixel.  The new "mirror"
is at the center of the edge pixels.  This avoids edge pixel
duplication.

Change-Id: I7f35fff5eda056bc58554671de11ed23bfc9c0ee