Avoid /./ in mpi test executable path
In commit 5b9cbfd0 (Fix mpi test executable path for Makefile
generators, 2010-12-21) we replaced use of ${CTEST_CONFIGURATION_TYPE}
with just "." for single-configuration generators.  That leads to paths
ending in "bin/./pvtk" which confuses the Tcl interpreter search path
rendering it unable to load the vtk Tcl package.  TestPProbe-image fails
as a result.

Instead append /${CTEST_CONFIGURATION_TYPE} to the path for multi-config
generators and nothing for single-config generators.

Change-Id: I60d993617f17a5c6a92948040757182aae5a32d7