vtkAxisActor2D tick distance was wrong in RulerMode
For the axis wLength of 101.0 and a rulerDistance of 50.0, the number of
ticks should be 4: 0, 50, 100 and 101.
When converting to viewport coordinates, if the axis length is 10.1,
the tick distance should then be 5.0.
Note that the number of ticks must contain the start and end ticks, so
they can be skipped in the for loop that adds tick points in the polyline.
If the last tick is confounded with the end tick mark, don't add the end
tick to numTicks.

Change-Id: I6713e7a0d8f90c7887c61fa7a1a64910c0555764