ENH: Make python initialize vtkTuple after construction.
In C++, the vtkTuple class and its subclasses no longer initialize
their contents in the default constructor.  In order to maintain
backwards compatibility within VTK/Python and to make the behavior
more pythonic, the wrappers now perform the initialization.

The behavior is not completely backwards-compatibile, because the
vtkColor4ub class used to initialize to [0,0,0,255] and now it
initializes to [0,0,0,0], ditto for vtkColor4v and vtkColor4d.

Change-Id: I056992dfb47b8ea1c0927c7ecc621a3e0649932b