Fixed bug where vtkImageReslice was not setting WHOLE_EXTENT properly.
When GenerateStencilOutput is false, vtkImageReslice was not
changing the WHOLE_EXTENT for the stencil output (output port 1).
However, by default, the executive copies the input WHOLE_EXTENT
to all output ports, which is wrong in this case. Added code
that removes WHOLE_EXTENT, SPACING and ORIGIN from second output
when GenerateStencilOutput is false.