ENH: Improvements to the scatter plot matrix.
Animations have been vastly improved, this has been achieved by storing
a map of column names with axis ranges. The same axis range is set on
all of the scatter plots (small, main and xyz) so that it is easier to
map between them. The animations now end up in the exact same
orientation and the 3D effect seems to be much more effective when
animating. For now turning off the axis rendering in 3D, I will revisit
that along with some of the rendering code.

Change-Id: I467a4e712ee8a35c1ff2824afaece8f2177fcfa1