ENH: Separate interactors from rendering.
The vtkInteractorStyleSwitch provides a good default for the
vtkRenderWindowInteractor, but it also introduces a dependency of the
rendering classes on the interaction classes. The new
vtkInteractorStyleSwitchBase breaks that dependency, so that an empty
interactor would be constructed if the interactor classes are not
available, but the traditional switch default is preserved if the
interaction module is linked.

This commit temorarily overrides in the vtkGraphicsFactory class, but
this will be moved to the general factory override logic in modular VTK.

Change-Id: If6d73f8796eb6315c7dabd61b40abe5a6eb598ad