ENH: Reduce the number of test cases run
The test was originally testing the integrity of ghost-data generation and ghost
array marking functionality for different number of partitions and number of ghost
level requests. Theses tests, albeit useful to gain confidence in the code, they were
significantly increase the runtime for dashboards which in some cases would fail the
test due to timeout. Changed the test so that it checks a single non-trivial test case
that should cover most of the functionality: 256 number of partitions/blocks and 3 levels
of ghost layers. Also removed any I/O that was done for debugging.

Change-Id: I960bca18aa9c8fd32d37b2546a7b2ebaefee85e8