Fixing vtkContextActor to work with tiled displays.
vtkContextActor wasn't using correct arithmetic to determine the scale/offset
when using with tiled displays. The code failed when the renderer's viewport was
not covering the entire display. This fixes that.

Adding a new internal class vtkViewportSpecification that may be a good
candidate to expose and be made public since we do similar arithmetic in many
places including ParaView.

Change-Id: Ia55e44298cdf54fdced943755755963176a78410