COMP: Fix type, var, loop warnings in vtkStructuredGridConnectivity.
In PrintSelf, a for() loop had no body.  In CreateGhostedMaskArrays,
the variable "p" caused unitialialized variable warnings on gcc 4.0.
In vtkPStructuredGridConnectivity::PostReceives, an "int" var was
used to hold a std::vector size.  I changed this to "unsigned int"
for consistency with the rest of the file, but in actuality most
occurrences of "unsigned int" in this file should be "size_t".

Change-Id: I4ace7d7531fc348134b08db197c5c2805f2e1a3a