Moves necessary after FiltersHybrid changes
The FiltersHybrid module needs RenderingCore for vtkCamera. The
GeovisCore module has a hard dependency on RenderingOpenGL for the
vtkOpenGLHardwareSupprt class, and other things, this also means it must
be removed from the StandAlone group due to the GL dependency.

The InteractionWidgets module had a few dependencies that were not
clear, and has an OpenGL dependency for one of the widgets that looks
like it can be factored out.

Change-Id: Ibe14692c952f76bdf9cd8b13a25b63c158598693