Create an include with all instantiators for Python
This brings back the old behavior when instantiators are turned on,
bringing in all VTK symbols at link time of vtkpython, rather than when
the modules are imported by Python in the interpreter.

Disabled the include for now as in my testing it consistently slowed
down test execution. This should probably be removed in the future, but
is useful for now in case others would like to do further testing.

Change-Id: I476d96c5c13c5e45efa1c3c86595afb70c104f9d