BUG: Made void GetPoint(vtkIdType, double x[3]) threadsafe
The method 'void GetPoint(vtkIdType, double x[3])', which is supposed
to be threadsafe, called the un-threadsafe method
'double * GetPoint(vtkIdType)'. I reversed the relationship of the
methods to make the former method actually threadsafe. Also, I replaced
the static variable that held the point location in the non-threadsafe
method with a member variable.

Change-Id: I44b79521a265aefdf9b54794f95be63f6c89c256