COMP: No buffer lookahead space for wrapper-parser include directives.
The wrapper's C++ preprocessor has its own code for parsing header
files.  Its file buffer did not have any reserve for lookahead chars.
Lookahead is done whenever a backslash appears in the file, so when
a backslash was the last character in the buffer, an invalid read
occurred.  This patch reserves two chars at the end of the buffer
for lookahead.

Change-Id: I9456e3e8b8e587e5fd92947e9a73acfd39ff3bae