vtkAutoInit: Do not use VTK_ABI_IMPORT directly
The VTK_ABI_IMPORT macro tells us *how* to import a symbol known to be
in a shared library, not *whether* it is in one.  It is not meant for
direct use and doing so tries to dllimport symbols from static libraries
on Windows.

Teach VTK_AUTOINIT to declare <mod>_AutoInit_(Construct|Destruct)
without any export/import markup.  The symbols are already exported
where they are defined.  Since they are not data symbols the linker will
resolve them in either a static library or dll import library
automatically.

Change-Id: I9939c6ed5946e8413fb6a9f3987b159e0d943530