Store the viewport and tile size in the device
Store the total viewport size and the tile size/location in the device.
This reduces the need to recompute these values, and allows the context
device to take care of translating clipping coordinates correctly. This
also removed the need for the vtkContextClip device to use the scene
transform for the clip coordinates.

Change-Id: Iebd9ba58607f5221d92b4b6653a11f9fe2e9f8b4