Remove '--' prefix from module summary format
For modules with multiple dependents we list each on a new line.
Previously we added a '--' prefix to each such line to match the
message(STATUS ...) output produced by CMake on the first line.
However, in cmake-gui there is no prefix so '--' makes the dependents
stand out more than the modules.  Remove '--' to fix that.

cmake-gui before:

  * vtkCommonCore, needed by 4 modules:
 --      vtkCommonMath
 --      vtkCommonSystem
 --      vtkIOExport
 --      vtkIOImport
  * vtkDICOMParser, needed by vtkIOImage.

cmake-gui after:

  * vtkCommonCore, needed by 4 modules:
         vtkCommonMath
         vtkCommonSystem
         vtkIOExport
         vtkIOImport
  * vtkDICOMParser, needed by vtkIOImage.

Change-Id: I4dd287dd72b68dc2b35b79e6876a87664e2fab65