Add API to allow users to set desired output point precision
This adds new API to a few of the vtkAlgorithm sub-classes which
allows the user to decide what level of floating-point precision
they want for the vtkPoint's object in the output data set. By
default these algorithms will output points with a precision equal
to that of the input data set. Alternatively the user may instruct
the algorithm to always use single-precision or double-precision.

Change-Id: I1c6b8ca854ff8da1bae5415e2d4c33fbfbaf1153