Refactor vtkPNrrReader into parallel and non parallel version
The idea here is to have a base class vtkNrrdReader that can be
used without MPI and contains no #ifdefs. The subclass vtkPNrrdReader
is an MPI version. The object factory overrides are then used to
create the correct version depending on whether you link to the
parallel module. In order todo this the MPI code from vtkMPIImageReader
has to be pushed down into vtkPNrrdReader. vtkMPIImageReader has been
removed.

Change-Id: I4839c72711333ddbd8a79056391299a8c420ed21