BUG: Make the Python tests do image comparisons
A few changes to account for pipeline changes made in master. This
causes a test failure when the baseline image cannot be found, and also
ensures that the image compare is performed. I verified by changing one
of the baselines - it may be an idea to add an expected fail for this
type of thing.

ENH: Remove reference to Dart

Dart is a legacy dashboard system. Changed internal proc names to use
CDash versus Dart.

Change-Id: I8de94fe03166ee475951f373f327408da1c58d35