ENH: Stripping partially overlapping ghost-layers
Added functionality for stripping out partially overlapping
ghost layers. For each grid a ghost vector is computed that
encodes for each of the 6 sides [imin,imax,jmin,jmax,kmin,kmax]
the number of ghost layers to strip out,i.e., the number of
cells that partially overlap with a lower resolution cell. Then,
from the ghosted AMR dataset, a non-ghosted instance is created.

Change-Id: I9b2c200f948aca0daf82c14844baf2a17d1db838