BUGFIX: Fix issue with blanked cells
Fixed issue due to blanked cells. When a cell is blanked
an empty cell is returned by vtkUniformGrid, consequently,
the cell-centers were not computed correctly leading to a
test failure since the computed cell centers did not match
the centers that are attached as fields on the grid. Also,
made the test always create partially overlapping ghost layers.

Change-Id: Ic75743446154e36cabf1336dc065785136635b93