COMP: Move PythonCore to Python, make it a module again.
In VTK 5.10, the vtkPythonCore library was built in the Wrapping/Python
directory.  Modularization renamed the library to vtkWrappingPythonCore
and built it in Wrapping/PythonCore.  Soon afterwards the library was
named back to vtkPythonCore (its original name), and it became just a
library (not a modular-style module).

This patch moves the code back to Wrapping/Python (its original location),
renames vtkPythonCore to its a new name vtkWrappingPython, and makes the
library into a modular-style module again.

Change-Id: Ie41c2bfbe871b085d046e98af9a04a81ed535b37