vtkRenderWindow: ability to use the same offsets at focal depth rendering.
The random offsets in vtkRenderWindow::DoFDRender() add a very annoying
vibration to focal depth rendering. Let's say we would like to create a simple
video of moving the focal point smoothly from a frontal object to an object in
the background. At the beginning the frontal object is sharp, the background is
blurred, while at the end it is swapped. It should be easily done in VTK by
focal depth rendering and changing only the distance of the focal point. But
due to the fact that for each rendered frames new random offsets are calculated
for radius and angle ( in vtkRenderWindow::DoFDRender() ), the result is a
vibrating camera.

More simply: when focal depth rendering is turned on, each rendering yields
a different image, even if camera and scene do not change at all.

vtkRenderWindow is now modified: a new ivar is added to be able to select
constant offsets. Constant offsets are calculated when FDFrames changes. We can
decide to use constant offsets or not. Default is not using.

Change-Id: Id2bf0e75d4ca0f4b90f03efa08fa0d3a623011b8