vtkPython.h: Do not include 'sal.h' directly
In our "debug VTK with release Python" hack we include system header
files normally included by Python.h but while _DEBUG is still defined.
Drop inclusion of 'sal.h' because Python.h does not include it directly
and it does not exist on VS 7.1.  In versions of VS that have the header
it is indirectly included through other system headers anyway.

Change-Id: I1084006931441150b48ea16aea0f899d8d2d2683