Allow modules to not participate in hierarchy wrapping.
Now that ParaView's VisItBridge is a module it is being wrapped
automatically, and the automatic wrapping does the hierachy traversal.
As a kit the wrapping did not but the wrapped lib worked well enough.
The problem is that hierarchy wrapping brings in files from VisIt
that break the wrapper generators. This change makes it work more
like it used to.

Change-Id: I17c138e37f45b52775e8d0fbd02f417f750a663f