fix infovis tests
Tests segfault when vtkTextMapper::New() returned a NULL pointer.

vtkTextMapper it turns out is an interface class for which
vtkOpenGLFreeTypeTextMapper is, for now, the only implementation. There could
be other implementations, so making a hard module dependency which cmake would
resolve automatically isn't the right solution. Instead, unfortunately you
have to manually provide a dependency to Rendering/FreeTypeOpenGL.

Change-Id: I0cf1e81bd14e226d09df07b2fcddae3f3df1404e