BUG: Fix loop in vtkHardwareSelector::GetPixelInformation
When vtkHardwareSelector::GetPixelInformation was called for the pixel
{0, 0}, the function would effectively freeze while stuck in a loop.

The bug was introduced by commit dd4f84d805027c0cec8dd5ebfbed54f366e3798d.

This commit fixes the bug by reverting the previous commit, and also
introduces logic to reduce the number of unnecessary recursive calls
when near the edge of the screen.

Change-Id: If711026da900e80e0419c79bec0ddf26c8f646cf