fix more warnings
I previously overlooked a couple unused-but-not-set-variable
warnings in the vtkNewickTreeReader because my local version
of GCC isn't new enough to detect these.  Per Dave Demarle's
request, I also fixed a couple signed/unsigned comparison
warnings in vtkBoostDividedEdgeBundling.cxx.  There still
remains an array-bounds warnings in this file, but I'm not
comfortable attempting to fix it at this time.

Change-Id: I9d760f2c712776275e1cb566f0de013722601b3f