BUG #13418. Adding support to render AMR composite ids for selection.
Extending vtkHardwareSelectionPolyDataPainter to add API to provide an array to
use for composite ids. Geometry filters can use this mechanism to provide
original composite ids for an amr dataset that has been converted to multiblock
of poly data. This is needed since composite indices on the rendered multiblock
are not same as those on the AMR dataset since AMR uses a 0-based index starting
with its first child.

Also fixing vtkHardwareSelector to handle composite index 0. Since 0 is now a
valid composite index and not merely "everything" as was the case before.

Change-Id: I97318cb340f71384b63e4ec62b97e7403b5aba31