Enable vtkTree I/O with R and multiple inputs in vtkRCalculaterFilter.
1: Implemented the conversion between a
R phylo tree and a VTKTree.

A R "phylo" tree is a list of 5 elements, including:
edge, Nnode,edge.length, tip.label, node.label.

The indexing convension of the vertices of a tree is
different between a "phylo" tree and the vtk tree.

2: Modified the vtkRCalculaterFilter so that it can take
in multipe input data types.For example,now you could add
a table, a tree and an array to the input port 0. The filter
will convert all vtk data inputs into R variables before
running the RScripts. For now, the ouput number remains to be
one. In the future, further modifications are needed to allow
multiple outputs.

Change-Id: Iba030e0d207c204aacec1e3058fb60bc954ffea0